Conflict resolved

+++ src/SalesManager.java
@@ -1,13 +1,13 @@
 public class SalesManager {
-    protected int[] sales;
+    protected long[] sales;
 
-    public SalesManager(int[] sales) {
+    public SalesManager(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
@@ -16,9 +16,9 @@ public class SalesManager {
     }
 
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Long.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -26,9 +26,9 @@ public class SalesManager {
         return min;
     }
 
-    public int avgTrimmed() {
-        int sum = 0;
-        for (int sale : sales) {
+    public long avgTrimmed() {
+        long sum = 0;
+        for (long sale : sales) {
             sum +=sale;
         }
         return (sum - min() - max()) / sales.length - 2);
